requirements.txt
streamlit
tensorflow==2.15.0
pillow
numpy
scikit-image
h5py
plotly